requirements-full.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
cloudflare==2.11.1
deepdiff==6.3.0
idna==3.4
jsonlines==3.1.0
logging==0.4.9.6
ordered-set==4.1.0
pi-hole-api==0.1.5
PyYAML==6.0
requests==2.29.0
soupsieve==2.4.1
urllib3==1.26.15
